Directory cleanup and creation of model.py
